Move to PVHVM 14.04 image
The Rax public cloud 14.04 PV image is missing the vhost_net module
which is required by OSA. vhost_net is not a new requirement, so this
problem may have been caused by an update to the 14.04 PV image.

This commit switches to using the Rax Pub
cloud PVHVM 14.04 image for heat multinode builds.

Connects rcbops/u-suk-dev#273